+++ axelor-business-project/src/main/java/com/axelor/apps/businessproject/service/InvoicingProjectService.java
@@ -56,14 +56,17 @@ import com.axelor.exception.AxelorException;
 import com.axelor.exception.db.repo.TraceBackRepository;
 import com.axelor.i18n.I18n;
 import com.axelor.inject.Beans;
+import com.axelor.team.db.TeamTask;
 import com.axelor.team.db.repo.TeamTaskRepository;
 import com.google.inject.Inject;
 import com.google.inject.persist.Transactional;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
 public class InvoicingProjectService {
@@ -73,6 +76,8 @@ public class InvoicingProjectService {
   @Inject protected ExpenseService expenseService;
 
   @Inject protected PartnerService partnerService;
+
+  @Inject protected TeamTaskBusinessService teamTaskBusinessService;
 
   @Inject protected InvoicingProjectRepository invoicingProjectRepo;
 
@@ -139,6 +144,7 @@ public class InvoicingProjectService {
         new ArrayList<PurchaseOrderLine>(folder.getPurchaseOrderLineSet());
     List<TimesheetLine> timesheetLineList = new ArrayList<TimesheetLine>(folder.getLogTimesSet());
     List<ExpenseLine> expenseLineList = new ArrayList<ExpenseLine>(folder.getExpenseLineSet());
+    List<TeamTask> teamTaskList = new ArrayList<TeamTask>(folder.getTeamTaskSet());
 
     List<InvoiceLine> invoiceLineList = new ArrayList<InvoiceLine>();
     invoiceLineList.addAll(
@@ -153,6 +159,9 @@ public class InvoicingProjectService {
     invoiceLineList.addAll(
         expenseService.createInvoiceLines(
             invoice, expenseLineList, folder.getExpenseLineSetPrioritySelect()));
+    invoiceLineList.addAll(
+        teamTaskBusinessService.createInvoiceLines(
+            invoice, teamTaskList, folder.getTeamTaskSetPrioritySelect()));
 
     Collections.sort(invoiceLineList, new InvoiceLineComparator());
 
@@ -280,6 +289,9 @@ public class InvoicingProjectService {
     for (ExpenseLine expenseLine : invoicingProject.getExpenseLineSet()) {
       expenseLine.setInvoiced(true);
     }
+    for (TeamTask teamTask : invoicingProject.getTeamTaskSet()) {
+      teamTask.setInvoiced(true);
+    }
     for (Project project : invoicingProject.getProjectSet()) {
       project.setInvoiced(true);
     }
@@ -307,107 +319,110 @@ public class InvoicingProjectService {
   public void fillLines(InvoicingProject invoicingProject, Project project) {
     if (project.getProjInvTypeSelect() == ProjectRepository.INVOICING_TYPE_FLAT_RATE
         || project.getProjInvTypeSelect() == ProjectRepository.INVOICING_TYPE_TIME_BASED) {
+
+      String commonQuery =
+          "self.project = :project AND self.toInvoice = true AND self.invoiced = false";
+
+      StringBuilder solQueryBuilder = new StringBuilder(commonQuery);
+      solQueryBuilder.append(
+          " AND (self.saleOrder.statusSelect = :statusConfirmed OR self.saleOrder.statusSelect = :statusCompleted)");
+
+      Map<String, Object> solQueryMap = new HashMap<>();
+      solQueryMap.put("project", project);
+      solQueryMap.put("statusConfirmed", SaleOrderRepository.STATUS_ORDER_CONFIRMED);
+      solQueryMap.put("statusCompleted", SaleOrderRepository.STATUS_ORDER_COMPLETED);
+
+      StringBuilder polQueryBuilder = new StringBuilder(commonQuery);
+      polQueryBuilder.append(
+          " AND (self.purchaseOrder.statusSelect = 3 OR self.purchaseOrder.statusSelect = 4)");
+
+      Map<String, Object> polQueryMap = new HashMap<>();
+      polQueryMap.put("project", project);
+
+      StringBuilder logTimesQueryBuilder = new StringBuilder(commonQuery);
+      logTimesQueryBuilder.append(" AND self.timesheet.statusSelect = :timesheetStatus");
+
+      Map<String, Object> logTimesQueryMap = new HashMap<>();
+      logTimesQueryMap.put("project", project);
+      logTimesQueryMap.put("timesheetStatus", TimesheetRepository.STATUS_VALIDATED);
+
+      StringBuilder expenseLineQueryBuilder = new StringBuilder(commonQuery);
+      expenseLineQueryBuilder.append(
+          " AND (self.expense.statusSelect = :statusValidated OR self.expense.statusSelect = :statusReimbursed)");
+
+      Map<String, Object> expenseLineQueryMap = new HashMap<>();
+      expenseLineQueryMap.put("project", project);
+      expenseLineQueryMap.put("statusValidated", ExpenseRepository.STATUS_VALIDATED);
+      expenseLineQueryMap.put("statusReimbursed", ExpenseRepository.STATUS_REIMBURSED);
+
+      StringBuilder taskQueryBuilder = new StringBuilder(commonQuery);
+      taskQueryBuilder.append(" AND self.status = 'closed'");
+
+      Map<String, Object> taskQueryMap = new HashMap<>();
+      taskQueryMap.put("project", project);
+
       if (invoicingProject.getDeadlineDate() != null) {
-        invoicingProject
-            .getSaleOrderLineSet()
-            .addAll(
-                Beans.get(SaleOrderLineRepository.class)
-                    .all()
-                    .filter(
-                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.saleOrder.creationDate < ?2"
-                            + " AND (self.saleOrder.statusSelect = ?3 OR self.saleOrder.statusSelect = ?4)",
-                        project,
-                        invoicingProject.getDeadlineDate(),
-                        SaleOrderRepository.STATUS_ORDER_CONFIRMED,
-                        SaleOrderRepository.STATUS_ORDER_COMPLETED)
-                    .fetch());
+        solQueryBuilder.append(" AND self.saleOrder.creationDate < :deadlineDate");
+        solQueryMap.put("deadlineDate", invoicingProject.getDeadlineDate());
 
-        invoicingProject
-            .getPurchaseOrderLineSet()
-            .addAll(
-                Beans.get(PurchaseOrderLineRepository.class)
-                    .all()
-                    .filter(
-                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.purchaseOrder.orderDate < ?2"
-                            + " AND (self.purchaseOrder.statusSelect = 3 OR self.purchaseOrder.statusSelect = 4)",
-                        project,
-                        invoicingProject.getDeadlineDate())
-                    .fetch());
+        polQueryBuilder.append(" AND self.purchaseOrder.orderDate < :deadlineDate");
+        polQueryMap.put("deadlineDate", invoicingProject.getDeadlineDate());
 
-        invoicingProject
-            .getLogTimesSet()
-            .addAll(
-                Beans.get(TimesheetLineRepository.class)
-                    .all()
-                    .filter(
-                        "self.timesheet.statusSelect = ?1 AND self.project = ?2 AND self.toInvoice = true AND self.invoiced = false AND self.date < ?3",
-                        TimesheetRepository.STATUS_VALIDATED,
-                        project,
-                        invoicingProject.getDeadlineDate())
-                    .fetch());
+        logTimesQueryBuilder.append(" AND self.date < :deadlineDate");
+        logTimesQueryMap.put("deadlineDate", invoicingProject.getDeadlineDate());
 
-        invoicingProject
-            .getExpenseLineSet()
-            .addAll(
-                Beans.get(ExpenseLineRepository.class)
-                    .all()
-                    .filter(
-                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.expenseDate < ?2"
-                            + " AND (self.expense.statusSelect = ?3 OR self.expense.statusSelect = ?4",
-                        project,
-                        invoicingProject.getDeadlineDate(),
-                        ExpenseRepository.STATUS_VALIDATED,
-                        ExpenseRepository.STATUS_REIMBURSED)
-                    .fetch());
-      } else {
-        invoicingProject
-            .getSaleOrderLineSet()
-            .addAll(
-                Beans.get(SaleOrderLineRepository.class)
-                    .all()
-                    .filter(
-                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false"
-                            + " AND (self.saleOrder.statusSelect = ?2 OR self.saleOrder.statusSelect = ?3)",
-                        project,
-                        SaleOrderRepository.STATUS_ORDER_CONFIRMED,
-                        SaleOrderRepository.STATUS_ORDER_COMPLETED)
-                    .fetch());
+        expenseLineQueryBuilder.append(" AND self.expenseDate < :deadlineDate");
+        expenseLineQueryMap.put("deadlineDate", invoicingProject.getDeadlineDate());
 
-        invoicingProject
-            .getPurchaseOrderLineSet()
-            .addAll(
-                Beans.get(PurchaseOrderLineRepository.class)
-                    .all()
-                    .filter(
-                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false"
-                            + " AND (self.purchaseOrder.statusSelect = 3 OR self.purchaseOrder.statusSelect = 4)",
-                        project)
-                    .fetch());
-
-        invoicingProject
-            .getLogTimesSet()
-            .addAll(
-                Beans.get(TimesheetLineRepository.class)
-                    .all()
-                    .filter(
-                        "self.timesheet.statusSelect = ?1 AND self.project = ?2 AND self.toInvoice = true AND self.invoiced = false",
-                        TimesheetRepository.STATUS_VALIDATED,
-                        project)
-                    .fetch());
-
-        invoicingProject
-            .getExpenseLineSet()
-            .addAll(
-                Beans.get(ExpenseLineRepository.class)
-                    .all()
-                    .filter(
-                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false"
-                            + " AND (self.expense.statusSelect = ?2 OR self.expense.statusSelect = ?3)",
-                        project,
-                        ExpenseRepository.STATUS_VALIDATED,
-                        ExpenseRepository.STATUS_REIMBURSED)
-                    .fetch());
+        taskQueryBuilder.append(" AND self.taskDeadline < :deadlineDate");
+        taskQueryMap.put("deadlineDate", invoicingProject.getDeadlineDate());
       }
+
+      invoicingProject
+          .getSaleOrderLineSet()
+          .addAll(
+              Beans.get(SaleOrderLineRepository.class)
+                  .all()
+                  .filter(solQueryBuilder.toString())
+                  .bind(solQueryMap)
+                  .fetch());
+
+      invoicingProject
+          .getPurchaseOrderLineSet()
+          .addAll(
+              Beans.get(PurchaseOrderLineRepository.class)
+                  .all()
+                  .filter(polQueryBuilder.toString())
+                  .bind(polQueryMap)
+                  .fetch());
+
+      invoicingProject
+          .getLogTimesSet()
+          .addAll(
+              Beans.get(TimesheetLineRepository.class)
+                  .all()
+                  .filter(logTimesQueryBuilder.toString())
+                  .bind(logTimesQueryMap)
+                  .fetch());
+
+      invoicingProject
+          .getExpenseLineSet()
+          .addAll(
+              Beans.get(ExpenseLineRepository.class)
+                  .all()
+                  .filter(expenseLineQueryBuilder.toString())
+                  .bind(expenseLineQueryMap)
+                  .fetch());
+
+      invoicingProject
+          .getTeamTaskSet()
+          .addAll(
+              Beans.get(TeamTaskRepository.class)
+                  .all()
+                  .filter(taskQueryBuilder.toString())
+                  .bind(taskQueryMap)
+                  .fetch());
+
       if (project.getProjInvTypeSelect() == ProjectRepository.INVOICING_TYPE_FLAT_RATE
           && !project.getInvoiced()) invoicingProject.addProjectSetItem(project);
     }
@@ -420,6 +435,7 @@ public class InvoicingProjectService {
     invoicingProject.setLogTimesSet(new HashSet<TimesheetLine>());
     invoicingProject.setExpenseLineSet(new HashSet<ExpenseLine>());
     invoicingProject.setProjectSet(new HashSet<Project>());
+    invoicingProject.setTeamTaskSet(new HashSet<TeamTask>());
   }
 
   public Company getRootCompany(Project project) {